#8 러닝세션 스크린 구현

+++ data/src/main/java/com/dpm/sixpack/data/repository/SensorRepositoryImpl.kt
@@ -12,7 +12,7 @@ import javax.inject.Inject
 class SensorRepositoryImpl @Inject constructor(
     private val sensorDataSource: SensorDataSource,
 ) : SensorRepository {
-    override fun getTotalStep(): Flow<DoRunResult<Long>> =
+    override fun getTotalStep(): Flow<DoRunResult<Int>> =
         sensorDataSource
             .getTotalStepsFlow()
             .map { stepCount ->
