+++ domain/src/main/java/com/dpm/sixpack/domain/repository/SensorRepository.kt
@@ -4,5 +4,5 @@ import com.dpm.sixpack.domain.util.DoRunResult
 import kotlinx.coroutines.flow.Flow
 
 interface SensorRepository {
-    fun getTotalStep(): Flow<DoRunResult<Long>>
+    fun getTotalStep(): Flow<DoRunResult<Int>>
 }
